requirements.txt
playwright
pandas
openpyxl
pymongo
dnspython
python-dotenv